move SegmentationROI into its own module
