+++ tests/modules/segmentation/postprocess_utils/test_segmentation_roi.py
@@ -1,8 +1,9 @@
 import pytest
 import numpy as np
 from ophys_etl.modules.decrosstalk.ophys_plane import OphysROI
+from ophys_etl.modules.segmentation.postprocess_utils.roi_types import (
+    SegmentationROI)
 from ophys_etl.modules.segmentation.postprocess_utils.roi_merging import (
-    SegmentationROI,
     do_rois_abut,
     merge_segmentation_rois,
     _get_rings,
